+++ components/test_raftstore/src/node.rs
@@ -14,8 +14,7 @@ use raft::SnapshotStatus;
 use super::*;
 use engine::*;
 use engine_rocks::RocksEngine;
-use tikv::config::ConfigHandler;
-use tikv::config::{ConfigController, TiKvConfig};
+use tikv::config::{ConfigController, ConfigHandler, Module, TiKvConfig};
 use tikv::import::SSTImporter;
 use tikv::raftstore::coprocessor::config::SplitCheckConfigManager;
 use tikv::raftstore::coprocessor::CoprocessorHost;
@@ -226,14 +225,17 @@ impl Simulator for NodeCluster {
         );
         split_check_worker.start(split_check_runner).unwrap();
         cfg_controller.register(
-            "coprocessor",
+            Module::Coprocessor,
             Box::new(SplitCheckConfigManager(split_check_worker.scheduler())),
         );
 
         let mut raftstore_cfg = cfg.raft_store.clone();
         raftstore_cfg.validate().unwrap();
         let raft_store = Arc::new(VersionTrack::new(raftstore_cfg));
-        cfg_controller.register("raft_store", Box::new(RaftstoreConfigManager(raft_store)));
+        cfg_controller.register(
+            Module::Raftstore,
+            Box::new(RaftstoreConfigManager(raft_store)),
+        );
         let config_client = ConfigHandler::start(
             cfg.server.advertise_addr,
             cfg_controller,
