+++ src/server/lock_manager/deadlock.rs
@@ -289,8 +289,8 @@ pub enum Task {
     ChangeRole(Role),
     /// Change the ttl of DetectTable
     ChangeTTL(Duration),
-    /// Task only used for test
-    #[cfg(test)]
+    // Task only used for test
+    #[cfg(any(test, feature = "testexport"))]
     Validate(Box<dyn FnOnce(u64) + Send>),
     #[cfg(test)]
     GetRole(Box<dyn FnOnce(Role) + Send>),
@@ -307,7 +307,7 @@ impl Display for Task {
             Task::DetectRpc { .. } => write!(f, "Detect Rpc"),
             Task::ChangeRole(role) => write!(f, "ChangeRole {{ role: {:?} }}", role),
             Task::ChangeTTL(ttl) => write!(f, "ChangeTTL {{ ttl: {:?} }}", ttl),
-            #[cfg(test)]
+            #[cfg(any(test, feature = "testexport"))]
             Task::Validate(_) => write!(f, "Validate dead lock config"),
             #[cfg(test)]
             Task::GetRole(_) => write!(f, "Get role of the deadlock detector"),
@@ -365,7 +365,7 @@ impl Scheduler {
         self.notify_scheduler(Task::ChangeTTL(Duration::from_millis(t)));
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "testexport"))]
     pub fn validate(&self, f: Box<dyn FnOnce(u64) + Send>) {
         self.notify_scheduler(Task::Validate(f));
     }
@@ -867,7 +867,7 @@ where
             }
             Task::ChangeRole(role) => self.handle_change_role(role),
             Task::ChangeTTL(ttl) => self.handle_change_ttl(ttl),
-            #[cfg(test)]
+            #[cfg(any(test, feature = "testexport"))]
             Task::Validate(f) => f(self.inner.borrow().detect_table.ttl.as_millis() as u64),
             #[cfg(test)]
             Task::GetRole(f) => f(self.inner.borrow().role),
