+++ tests/integrations/config/dynamic/mod.rs
@@ -1,5 +1,6 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
 mod gc_worker;
+mod pessimistic_txn;
 mod raftstore;
 mod split_check;
