+++ tests/integrations/config/test_config_client.rs
@@ -277,7 +277,7 @@ fn test_dispatch_change() {
         let (version, cfg) = ConfigHandler::create(id.to_owned(), pd_client.clone(), cfg).unwrap();
         *mgr.0.lock().unwrap() = cfg.raft_store.clone();
         let mut controller = ConfigController::new(cfg, version);
-        controller.register("raft_store", Box::new(mgr.clone()));
+        controller.register(Module::Raftstore, Box::new(mgr.clone()));
         ConfigHandler::start(
             id.to_owned(),
             controller,
